Bug 1183145: Move platform-specific code from addrs.c into separate files. r=mjf
Differential Revision: https://phabricator.services.mozilla.com/D75947
